feat: Add comprehensive README.md
This commit introduces a detailed README.md file that provides in-depth information about the gotypesensequery package. The README covers features, usage instructions, functionality for different field types, human-readable date conversions, contributing guidelines, and licensing details.

Additionally, it includes badges for Go Report Card and a table of contents for better navigation. The goal is to make it easier for users and contributors to understand and use the gotypesensequery package effectively.

Closes #1